+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 -   Increase width of dropdown when selecting a curse mod file
 -   Add some basic filtering to not show files for a different loader
 -   Add a warning and button when Fabric API not installed for Fabric loader
+-   Show dependant mods when installing a mod to an instance